LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-window/location-samedocument-sameorigin-expected.txt



FAIL using location.href to navigate same-document targeting a same-origin window assert_equals: navigationType expected "push" but got "replace"
